Moved General to a module folder
